PVF: Refactor workers into separate crates, remove host dependency
