+++ node/core/pvf/common/src/pvf.rs
@@ -14,7 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::artifacts::ArtifactId;
 use parity_scale_codec::{Decode, Encode};
 use polkadot_parachain::primitives::ValidationCodeHash;
 use polkadot_primitives::ExecutorParams;
@@ -27,7 +26,7 @@ use std::{
 };
 
 #[cfg(test)]
-use crate::host::tests::TEST_PREPARATION_TIMEOUT;
+use crate::tests::TEST_PREPARATION_TIMEOUT;
 
 /// A struct that carries the exhaustive set of data to prepare an artifact out of plain
 /// Wasm binary
@@ -58,13 +57,8 @@ impl PvfPrepData {
 		Self { code, code_hash, executor_params, prep_timeout }
 	}
 
-	/// Returns artifact ID that corresponds to the PVF with given executor params
-	pub(crate) fn as_artifact_id(&self) -> ArtifactId {
-		ArtifactId::new(self.code_hash, self.executor_params.hash())
-	}
-
 	/// Returns validation code hash for the PVF
-	pub(crate) fn code_hash(&self) -> ValidationCodeHash {
+	pub fn code_hash(&self) -> ValidationCodeHash {
 		self.code_hash
 	}
 
@@ -84,15 +78,15 @@ impl PvfPrepData {
 	}
 
 	/// Creates a structure for tests
-	#[cfg(test)]
-	pub(crate) fn from_discriminator_and_timeout(num: u32, timeout: Duration) -> Self {
+	#[doc(hidden)]
+	pub fn from_discriminator_and_timeout(num: u32, timeout: Duration) -> Self {
 		let descriminator_buf = num.to_le_bytes().to_vec();
 		Self::from_code(descriminator_buf, ExecutorParams::default(), timeout)
 	}
 
-	#[cfg(test)]
-	pub(crate) fn from_discriminator(num: u32) -> Self {
-		Self::from_discriminator_and_timeout(num, TEST_PREPARATION_TIMEOUT)
+	#[doc(hidden)]
+	pub fn from_discriminator(num: u32) -> Self {
+		Self::from_discriminator_and_timeout(num, crate::tests::TEST_PREPARATION_TIMEOUT)
 	}
 }
