+++ node/core/pvf/prepare-worker/Cargo.toml
@@ -1,35 +1,26 @@
 [package]
-name = "polkadot-node-core-pvf-worker"
+name = "polkadot-node-core-pvf-prepare-worker"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
 
-[[bin]]
-name = "puppet_worker"
-path = "bin/puppet_worker.rs"
-
 [dependencies]
-assert_matches = "1.4.0"
-cpu-time = "1.0.0"
 futures = "0.3.21"
 gum = { package = "tracing-gum", path = "../../../gum" }
 libc = "0.2.139"
 rayon = "1.5.1"
-tempfile = "3.3.0"
 tikv-jemalloc-ctl = { version = "0.5.0", optional = true }
-tokio = "1.24.2"
+tokio = { version = "1.24.2", features = ["fs", "process"] }
 
 parity-scale-codec = { version = "3.4.0", default-features = false, features = ["derive"] }
 
-polkadot-node-core-pvf = { path = ".." }
+polkadot-node-core-pvf-common = { path = "../common" }
 polkadot-parachain = { path = "../../../../parachain" }
 polkadot-primitives = { path = "../../../../primitives" }
 
 sc-executor = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sc-executor-common = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sc-executor-wasmtime = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-externalities = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-io = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-maybe-compressed-blob = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-tracing = { git = "https://github.com/paritytech/substrate", branch = "master" }
@@ -37,13 +28,6 @@ sp-tracing = { git = "https://github.com/paritytech/substrate", branch = "master
 [target.'cfg(target_os = "linux")'.dependencies]
 tikv-jemalloc-ctl = "0.5.0"
 
-[build-dependencies]
-substrate-build-script-utils = { git = "https://github.com/paritytech/substrate", branch = "master" }
-
-[dev-dependencies]
-adder = { package = "test-parachain-adder", path = "../../../../parachain/test-parachains/adder" }
-halt = { package = "test-parachain-halt", path = "../../../../parachain/test-parachains/halt" }
-tempfile = "3.3.0"
-
 [features]
+builder = []
 jemalloc-allocator = ["dep:tikv-jemalloc-ctl"]
