+++ node/core/pvf/src/execute/worker_intf.rs
@@ -18,17 +18,20 @@
 
 use crate::{
 	artifacts::ArtifactPathId,
-	error::InternalValidationError,
-	worker_common::{
-		framed_recv, framed_send, path_to_bytes, spawn_with_program_path, IdleWorker, SpawnErr,
-		WorkerHandle, JOB_TIMEOUT_WALL_CLOCK_FACTOR,
+	worker_intf::{
+		path_to_bytes, spawn_with_program_path, IdleWorker, SpawnErr, WorkerHandle,
+		JOB_TIMEOUT_WALL_CLOCK_FACTOR,
 	},
 	LOG_TARGET,
 };
 use futures::FutureExt;
 use futures_timer::Delay;
 use parity_scale_codec::{Decode, Encode};
-
+use polkadot_node_core_pvf_common::{
+	error::InternalValidationError,
+	execute::{Handshake, Response},
+	framed_recv, framed_send,
+};
 use polkadot_parachain::primitives::ValidationResult;
 use polkadot_primitives::ExecutorParams;
 use std::{path::Path, time::Duration};
@@ -207,43 +210,4 @@ async fn recv_response(stream: &mut UnixStream) -> io::Result<Response> {
 			format!("execute pvf recv_response: decode error: {:?}", e),
 		)
 	})
-}
-
-/// The payload of the one-time handshake that is done when a worker process is created. Carries
-/// data from the host to the worker.
-#[derive(Encode, Decode)]
-pub struct Handshake {
-	/// The executor parameters.
-	pub executor_params: ExecutorParams,
-}
-
-/// The response from an execution job on the worker.
-#[derive(Encode, Decode)]
-pub enum Response {
-	/// The job completed successfully.
-	Ok {
-		/// The result of parachain validation.
-		result_descriptor: ValidationResult,
-		/// The amount of CPU time taken by the job.
-		duration: Duration,
-	},
-	/// The candidate is invalid.
-	InvalidCandidate(String),
-	/// The job timed out.
-	TimedOut,
-	/// An unexpected panic has occurred in the execution worker.
-	Panic(String),
-	/// Some internal error occurred.
-	InternalError(InternalValidationError),
-}
-
-impl Response {
-	/// Creates an invalid response from a context `ctx` and a message `msg` (which can be empty).
-	pub fn format_invalid(ctx: &'static str, msg: &str) -> Self {
-		if msg.is_empty() {
-			Self::InvalidCandidate(ctx.to_string())
-		} else {
-			Self::InvalidCandidate(format!("{}: {}", ctx, msg))
-		}
-	}
 }
