+++ node/core/pvf/src/testing.rs
@@ -19,34 +19,10 @@
 //! N.B. This is not guarded with some feature flag. Overexposing items here may affect the final
 //!      artifact even for production builds.
 
+#[doc(hidden)]
+pub use crate::worker_intf::{spawn_with_program_path, SpawnErr};
+
 use polkadot_primitives::ExecutorParams;
-
-/// A function that emulates the stitches together behaviors of the preparation and the execution
-/// worker in a single synchronous function.
-pub fn validate_candidate(
-	code: &[u8],
-	params: &[u8],
-) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
-	use crate::executor_intf::{prepare, prevalidate, Executor};
-
-	let code = sp_maybe_compressed_blob::decompress(code, 10 * 1024 * 1024)
-		.expect("Decompressing code failed");
-
-	let blob = prevalidate(&code)?;
-	let artifact = prepare(blob, &ExecutorParams::default())?;
-	let tmpdir = tempfile::tempdir()?;
-	let artifact_path = tmpdir.path().join("blob");
-	std::fs::write(&artifact_path, &artifact)?;
-
-	let executor = Executor::new(ExecutorParams::default())?;
-	let result = unsafe {
-		// SAFETY: This is trivially safe since the artifact is obtained by calling `prepare`
-		//         and is written into a temporary directory in an unmodified state.
-		executor.execute(&artifact_path, params)?
-	};
-
-	Ok(result)
-}
 
 /// Use this macro to declare a `fn main() {}` that will check the arguments and dispatch them to
 /// the appropriate worker, making the executable that can be used for spawning workers.
@@ -54,9 +30,10 @@ pub fn validate_candidate(
 macro_rules! decl_puppet_worker_main {
 	() => {
 		fn main() {
-			$crate::sp_tracing::try_init_simple();
+			sp_tracing::try_init_simple();
 
 			let args = std::env::args().collect::<Vec<_>>();
+			gum::trace!(target: $crate::LOG_TARGET, ?args, "running puppet worker");
 			if args.len() < 3 {
 				panic!("wrong number of arguments");
 			}
@@ -81,13 +58,41 @@ macro_rules! decl_puppet_worker_main {
 					std::thread::sleep(std::time::Duration::from_secs(5));
 				},
 				"prepare-worker" => {
-					$crate::prepare_worker_entrypoint(&socket_path, version);
+					polkadot_node_core_pvf_prepare_worker::worker_entrypoint(&socket_path, version);
 				},
 				"execute-worker" => {
-					$crate::execute_worker_entrypoint(&socket_path, version);
+					polkadot_node_core_pvf_execute_worker::worker_entrypoint(&socket_path, version);
 				},
 				other => panic!("unknown subcommand: {}", other),
 			}
 		}
 	};
+}
+
+/// A function that emulates the stitches together behaviors of the preparation and the execution
+/// worker in a single synchronous function.
+pub fn validate_candidate(
+	code: &[u8],
+	params: &[u8],
+) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
+	use polkadot_node_core_pvf_execute_worker::Executor;
+	use polkadot_node_core_pvf_prepare_worker::{prepare, prevalidate};
+
+	let code = sp_maybe_compressed_blob::decompress(code, 10 * 1024 * 1024)
+		.expect("Decompressing code failed");
+
+	let blob = prevalidate(&code)?;
+	let artifact = prepare(blob, &ExecutorParams::default())?;
+	let tmpdir = tempfile::tempdir()?;
+	let artifact_path = tmpdir.path().join("blob");
+	std::fs::write(&artifact_path, &artifact)?;
+
+	let executor = Executor::new(ExecutorParams::default())?;
+	let result = unsafe {
+		// SAFETY: This is trivially safe since the artifact is obtained by calling `prepare`
+		//         and is written into a temporary directory in an unmodified state.
+		executor.execute(&artifact_path, params)?
+	};
+
+	Ok(result)
 }
