+++ parachain/test-parachains/undying/collator/bin/puppet_worker.rs
@@ -14,4 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
 
-polkadot_node_core_pvf_worker::decl_puppet_worker_main!();
+polkadot_node_core_pvf::decl_puppet_worker_main!();
